separate getFocusGutterExtension function
